auxiliary_analysis/requirements.txt
numpy
pyyaml
pillow
matplotlib
torch
torchvision
open_clip_torch
